Fix a bug in initialization
Inside the Cashmere initialization, Constellation.activate() was
called which resulted in executors starting activities too soon.  In
a distributed setting, it is possible that a node initializes faster
than another node.  If the fast node generates activities that can be
stolen by a slow node (that already did the Cashmere initialization,
but had not yet completed the application-specific initialization,
this could lead to errors.  In this application, Executor.get() could
be called before Executor.initialize().